chore: generate usage
